add event delegations

+++ src/main/resources/static/recruitment/training/js/training-details.js
@@ -97,7 +97,7 @@ $(document).ready(function(){
         })
     }
 
-    $('#editTraining').on('click', function(event){
+    $('#load-layout').on('click', '#editTraining',function(event){
         event.preventDefault();
         let jsonForm = convertFormToJson($("form").serializeArray());
         let trainingConfirmationType = JSON.parse($('#trainingConfirmationType').val());
@@ -192,37 +192,49 @@ $(document).ready(function(){
                         icon: "error"
                     });
                 });
+            } else {
+                location.reload();
             }
         });
 
     }
 
-    $('#note-body').on("click","a.delete-note",function(event){
+    $('#load-layout').on("click","a.delete-note",function(event){
         event.preventDefault();
         let url = event.target.href;
-        $.ajax({
-            type:"DELETE",
-            url: url
-        }).then(function(response){
-            swal({
-                 title: "Success!",
-                 text: "You deleted a note",
-                 icon: "success",
-                 timer: 2000
-             }).then(function(){
-                location.reload();
-             });
-        }).fail(function(error){
-            console.log(error);
-            swal({
-                title: "Error!",
-                text: "Could NOT remove note! \n" + error.responseJSON.message,
-                icon: "error"
-            });
-        });
+        swal({
+           title: "Are you sure?",
+           text: "Once deleted, you will not be able to recover this note!",
+           icon: "warning",
+           buttons: true,
+           dangerMode: true,
+         }).then((willDelete) => {
+            if(willDelete){
+                $.ajax({
+                    type:"DELETE",
+                    url: url
+                }).then(function(response){
+                    swal({
+                         title: "Success!",
+                         text: "You deleted a note",
+                         icon: "success",
+                         timer: 2000
+                     }).then(function(){
+                        location.reload();
+                     });
+                }).fail(function(error){
+                    console.log(error);
+                    swal({
+                        title: "Error!",
+                        text: "Could NOT remove note! \n" + error.responseJSON.message,
+                        icon: "error"
+                    });
+                });
+            }
+         });
      })
 
-     $('#deleteTraining').on('click', function(event){
+     $('#load-layout').on('click', '#deleteTraining',function(event){
          event.preventDefault();
          let trainingId = vm.trainingDto.id;
          swal({
@@ -233,7 +245,9 @@ $(document).ready(function(){
            dangerMode: true,
          })
          .then((willDelete) => {
-            deleteTraining(trainingId);
+            if(willDelete){
+                deleteTraining(trainingId);
+            }
          });
      });
 
@@ -260,12 +274,12 @@ $(document).ready(function(){
          });
      }
 
-     $('#showApplication').on('click', function(event){
+     $('#load-layout').on('click', '#showApplication',  function(event){
         event.preventDefault();
         window.location.href= "/recruitment/application/application-edit.html?applicationId=" + vm.trainingDto.application.id;
      });
 
-     $('#showInterview').on('click', function(event){
+     $('#load-layout').on('click', '#showInterview', function(event){
         event.preventDefault();
         window.location.href= "/recruitment/interview/interview-details.html?interviewId=" + vm.trainingDto.interview.id;
      });
